Remove src/__init__.py and fix many errors (#979)
Major changes:

- Adds type checking on a few more files.
- Expr.dtype's type is no longer Optional[DataType], because that caused type havoc. Therefore, the default is changed from None to DataType.AUTO when not set explicitly. The gtir_dtype_resolver and other visitors are updated to check for DataType.AUTO instead of None.
- Moves most of the code that was in eve.iterators to eve.concepts to work around a circular dependency after solving ClassVar parameters cannot include any type variables #565 using @egparedes's suggestion. This is a temporary workaround until we replace Eve with the new version.
- gt4py.stencil_object.ArgsInfo is made non-private and used in gt4py.dace_stencil_object to fix a type issue.
- Ignores typing in eve visitor contexts -- need to fix that in a follow-up PR.
- npir.HorizontalMask is created as a tuple of tuples, to avoid the Optional[AxisBound] in the common.HorizontalInterval type.
- Replaces gt_backend references with gt4py.backend.

Follows #978.

Resolves #565.
Resolves #992.
Resolves #1007.

+++ src/gtc/gtir_to_oir.py
@@ -84,7 +84,7 @@ class GTIRToOIR(NodeTranslator):
         )
 
     def visit_UnaryOp(self, node: gtir.UnaryOp) -> oir.UnaryOp:
-        return oir.UnaryOp(op=node.op, expr=self.visit(node.expr), loc=node.loc)
+        return oir.UnaryOp(op=node.op, expr=self.visit(node.expr), dtype=node.dtype, loc=node.loc)
 
     def visit_BinaryOp(self, node: gtir.BinaryOp) -> oir.BinaryOp:
         return oir.BinaryOp(
